refactor: move CLI tests into system and separate
